uhd: remove device time read during alignment
Recent UHD firmware for the USRP2/N210 replaces the MicroBlaze
with a slower ZPU in addition to changes to the control
transactions. The effect is less predictable reading of the
current time and Tx/Rx sample mis-alignment following
underruns.

After an underrun, this patch drops potentially stale packets
with a fixed interval instead of relying on reading the
current time from the device.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>