ARM: separate summing and non-summing NEON vector operations
For a vector multiply-accumulate operations, the result can be
returned as either a single summed complex value, or left in
vector form. For iterative use, leaving the result in vector
form gains some efficiency at the expense of some code
complexity in that the summing and non-summing versions must
be maintained.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>